requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.2.1=pypi_0
certifi=2024.2.2=py39h06a4308_0
charset-normalizer=2.1.1=pypi_0
contourpy=1.2.0=pypi_0
cpuonly=2.0=0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
faiss-cpu=1.8.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
flatbuffers=23.1.4=pypi_0
fonttools=4.49.0=pypi_0
freetype=2.12.1=h4a9f257_0
gast=0.4.0=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.51.1=pypi_0
h5py=3.7.0=pypi_0
idna=3.4=py39h06a4308_0
imageio=2.34.0=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=6.1.3=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py39h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
keras=2.11.0=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=15.0.6.1=pypi_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.8.3=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=pypi_0
numpy=1.24.1=pypi_0
numpy-base=1.26.4=py39hb5e798b_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=8.7.0.84=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.0=pypi_0
pandas=2.2.1=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=py39h06a4308_0
protobuf=3.19.6=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.1.2=pypi_0
python=3.9.15=h7a1cb2a_2
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.3.0=py3.9_cpu_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cpu
pytorchcv=0.0.67=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=py39h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.40.1=h5082296_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-io-gcs-filesystem=0.29.0=pypi_0
tensorrt=8.5.2.2=pypi_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.3.0=pypi_0
tifffile=2024.2.12=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.3.0=py39_cpu
torchtriton=2.3.0=py39
torchvision=0.18.0=py39_cpu
tqdm=4.66.2=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024.1=pypi_0
urllib3=1.26.14=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
